Create clean branch for new LevelFinder
